+++ packages/mattermost-redux/src/actions/users.test.js
@@ -13,6 +13,7 @@ import {RequestStatus} from '../constants';
 import TestHelper from 'mattermost-redux/test/test_helper';
 import configureStore from 'mattermost-redux/test/test_store';
 import deepFreeze from 'mattermost-redux/utils/deep_freeze';
+import {UserTypes} from 'mattermost-redux/action_types';
 
 const OK_RESPONSE = {status: 'OK'};
 
@@ -54,71 +55,6 @@ describe('Actions.Users', () => {
         assert.ok(profiles[user.id]);
     });
 
-    it('login', async () => {
-        const user = TestHelper.basicUser;
-
-        nock(Client4.getBaseRoute()).
-            post('/users/logout').
-            reply(200, OK_RESPONSE);
-
-        await TestHelper.basicClient4.logout();
-
-        TestHelper.mockLogin();
-
-        await Actions.login(user.email, user.password)(store.dispatch, store.getState);
-
-        const state = store.getState();
-        const loginRequest = state.requests.users.login;
-        const {currentUserId, profiles} = state.entities.users;
-        const preferences = state.entities.preferences.myPreferences;
-        const teamMembers = state.entities.teams.myMembers;
-        const serverVersion = state.entities.general.serverVersion;
-
-        if (loginRequest.status === RequestStatus.FAILURE) {
-            throw new Error(JSON.stringify(loginRequest.error));
-        }
-
-        assert.ok(currentUserId);
-        assert.ok(profiles);
-        assert.ok(profiles[currentUserId]);
-        assert.ok(Object.keys(preferences).length);
-        assert.ok(serverVersion);
-
-        Object.keys(teamMembers).forEach((id) => {
-            assert.ok(teamMembers[id].team_id);
-            assert.equal(teamMembers[id].user_id, currentUserId);
-        });
-    });
-
-    it('loginById', async () => {
-        const user = TestHelper.basicUser;
-
-        nock(Client4.getBaseRoute()).
-            post('/users/logout').
-            reply(200, OK_RESPONSE);
-
-        await TestHelper.basicClient4.logout();
-
-        TestHelper.mockLogin();
-
-        await Actions.loginById(user.id, 'password1')(store.dispatch, store.getState);
-
-        const state = store.getState();
-        const {currentUserId, profiles} = state.entities.users;
-        const preferences = state.entities.preferences.myPreferences;
-        const teamMembers = state.entities.teams.myMembers;
-
-        assert.ok(currentUserId);
-        assert.ok(profiles);
-        assert.ok(profiles[currentUserId]);
-        assert.ok(Object.keys(preferences).length);
-
-        Object.keys(teamMembers).forEach((id) => {
-            assert.ok(teamMembers[id].team_id);
-            assert.equal(teamMembers[id].user_id, currentUserId);
-        });
-    });
-
     it('getTermsOfService', async () => {
         const response = {
             create_at: 1537976679426,
@@ -137,13 +73,15 @@ describe('Actions.Users', () => {
     });
 
     it('updateMyTermsOfServiceStatus accept terms', async () => {
-        const user = TestHelper.basicUser;
         nock(Client4.getBaseRoute()).
             post('/users').
             reply(201, {...TestHelper.fakeUserWithId()});
 
         TestHelper.mockLogin();
-        await Actions.login(user.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         nock(Client4.getBaseRoute()).
             post('/users/me/terms_of_service').
@@ -161,13 +99,15 @@ describe('Actions.Users', () => {
     });
 
     it('updateMyTermsOfServiceStatus reject terms', async () => {
-        const user = TestHelper.basicUser;
         nock(Client4.getBaseRoute()).
             post('/users').
             reply(201, {...TestHelper.fakeUserWithId()});
 
         TestHelper.mockLogin();
-        await Actions.login(user.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         nock(Client4.getBaseRoute()).
             post('/users/me/terms_of_service').
@@ -652,7 +592,10 @@ describe('Actions.Users', () => {
         assert.strictEqual(sessions.length, 0);
 
         TestHelper.mockLogin();
-        await Actions.loginById(user.id, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         nock(Client4.getBaseRoute()).
             post('/users/login').
@@ -705,7 +648,10 @@ describe('Actions.Users', () => {
         assert.strictEqual(sessions.length, 0);
 
         TestHelper.mockLogin();
-        await Actions.loginById(user.id, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         nock(Client4.getBaseRoute()).
             post('/users/login').
@@ -839,7 +785,10 @@ describe('Actions.Users', () => {
 
     it('updateMe', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, TestHelper.basicUser.password)(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const state = store.getState();
         const currentUser = state.entities.users.profiles[state.entities.users.currentUserId];
@@ -887,7 +836,10 @@ describe('Actions.Users', () => {
 
     it('patchUser', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, TestHelper.basicUser.password)(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const state = store.getState();
         const currentUserId = state.entities.users.currentUserId;
@@ -932,7 +884,10 @@ describe('Actions.Users', () => {
 
     it('updateUserRoles', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -950,7 +905,10 @@ describe('Actions.Users', () => {
 
     it('updateUserMfa', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -968,7 +926,10 @@ describe('Actions.Users', () => {
 
     it('updateUserPassword', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const beforeTime = new Date().getTime();
         const currentUserId = store.getState().entities.users.currentUserId;
@@ -1060,7 +1021,10 @@ describe('Actions.Users', () => {
 
     it('uploadProfileImage', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const testImageData = fs.createReadStream('packages/mattermost-redux/test/assets/images/test.png');
 
@@ -1082,7 +1046,10 @@ describe('Actions.Users', () => {
 
     it('setDefaultProfileImage', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -1146,7 +1113,10 @@ describe('Actions.Users', () => {
     it('createUserAccessToken', (done) => {
         async function test() {
             TestHelper.mockLogin();
-            await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+            store.dispatch({
+                type: UserTypes.LOGIN_SUCCESS,
+            });
+            await Actions.loadMeREST()(store.dispatch, store.getState);
 
             const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -1174,7 +1144,10 @@ describe('Actions.Users', () => {
 
     it('getUserAccessToken', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -1207,7 +1180,10 @@ describe('Actions.Users', () => {
 
     it('getUserAccessTokens', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -1241,7 +1217,10 @@ describe('Actions.Users', () => {
 
     it('getUserAccessTokensForUser', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -1275,7 +1254,10 @@ describe('Actions.Users', () => {
 
     it('revokeUserAccessToken', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -1320,7 +1302,10 @@ describe('Actions.Users', () => {
 
     it('disableUserAccessToken', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -1372,7 +1357,10 @@ describe('Actions.Users', () => {
 
     it('enableUserAccessToken', async () => {
         TestHelper.mockLogin();
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
 
@@ -1423,7 +1411,11 @@ describe('Actions.Users', () => {
     });
 
     it('clearUserAccessTokens', async () => {
-        await Actions.login(TestHelper.basicUser.email, 'password1')(store.dispatch, store.getState);
+        TestHelper.mockLogin();
+        store.dispatch({
+            type: UserTypes.LOGIN_SUCCESS,
+        });
+        await Actions.loadMeREST()(store.dispatch, store.getState);
 
         const currentUserId = store.getState().entities.users.currentUserId;
